Move mark config to `src/mark.ts` and remove unnecessary namespace
