+++ src/compile/scale.ts
@@ -2,7 +2,6 @@ import * as log from '../log';
 
 import {SHARED_DOMAIN_OPS} from '../aggregate';
 import {COLUMN, ROW, X, Y, X2, Y2, SHAPE, SIZE, COLOR, OPACITY, TEXT, hasScale, Channel} from '../channel';
-import {Orient} from '../config';
 import {SOURCE, STACKED_SCALE} from '../data';
 import {DateTime, isDateTime, timestamp} from '../datetime';
 import {ChannelDefWithScale, FieldDef, field} from '../fielddef';
@@ -476,7 +475,7 @@ export function rangeMixins(scale: Scale, model: Model, channel: Channel): any {
         if (scaleConfig.barSizeRange !== undefined) {
           return {range: scaleConfig.barSizeRange};
         }
-        const dimension = markConfig.orient === Orient.HORIZONTAL ? Y : X;
+        const dimension = markConfig.orient === 'horizontal' ? Y : X;
         return {range: [markConfig.barThinSize, model.scale(dimension).bandSize]};
       } else if (mark === TEXTMARK) {
         return {range: scaleConfig.fontSizeRange };
