+++ src/config.ts
@@ -1,7 +1,8 @@
-import {ScaleConfig, defaultScaleConfig} from './scale';
 import {AxisConfig, defaultAxisConfig, defaultFacetAxisConfig} from './axis';
 import {LegendConfig, defaultLegendConfig} from './legend';
-import {StackOffset} from './stack';
+import {MarkConfig, defaultMarkConfig} from './mark';
+import {ScaleConfig, defaultScaleConfig} from './scale';
+
 
 export interface CellConfig {
   width?: number;
@@ -73,81 +74,7 @@ export const defaultFacetConfig: FacetConfig = {
   cell: defaultFacetCellConfig
 };
 
-export namespace FontWeight {
-    export const NORMAL: 'normal' = 'normal';
-    export const BOLD: 'bold' = 'bold';
-}
-export type FontWeight = typeof FontWeight.NORMAL | typeof FontWeight.BOLD;
-
-export namespace Shape {
-    export const CIRCLE: 'circle' = 'circle';
-    export const SQUARE: 'square' = 'square';
-    export const CROSS: 'cross' = 'cross';
-    export const DIAMOND: 'diamond' = 'diamond';
-    export const TRIANGLEUP: 'triangle-up' = 'triangle-up';
-    export const TRIANGLEDOWN: 'triangle-down' = 'triangle-down';
-}
-export type Shape = typeof Shape.CIRCLE | typeof Shape.SQUARE | typeof Shape.CROSS | typeof Shape.DIAMOND | typeof Shape.TRIANGLEUP | typeof Shape.TRIANGLEDOWN;
-
-export namespace Orient {
-    export const HORIZONTAL: 'horizontal' = 'horizontal';
-    export const VERTICAL: 'vertical' = 'vertical';
-}
-export type Orient = typeof Orient.HORIZONTAL | typeof Orient.VERTICAL;
-
-export namespace HorizontalAlign {
-    export const LEFT: 'left' = 'left';
-    export const RIGHT: 'right' = 'right';
-    export const CENTER: 'center' = 'center';
-}
-export type HorizontalAlign = typeof HorizontalAlign.LEFT | typeof HorizontalAlign.RIGHT | typeof HorizontalAlign.CENTER;
-
-export namespace VerticalAlign {
-    export const TOP: 'top' = 'top';
-    export const MIDDLE: 'middle' = 'middle';
-    export const BOTTOM: 'bottom' = 'bottom';
-}
-export type VerticalAlign = typeof VerticalAlign.TOP | typeof VerticalAlign.MIDDLE | typeof VerticalAlign.BOTTOM;
-
-export namespace FontStyle {
-    export const NORMAL: 'normal' = 'normal';
-    export const ITALIC: 'italic' = 'italic';
-}
-export type FontStyle = typeof FontStyle.NORMAL | typeof FontStyle.ITALIC;
-
-export namespace Interpolate {
-    /** piecewise linear segments, as in a polyline */
-    export const LINEAR: 'linear' = 'linear';
-    /** close the linear segments to form a polygon */
-    export const LINEAR_CLOSED: 'linear-closed' = 'linear-closed';
-    /** alternate between horizontal and vertical segments, as in a step function */
-    export const STEP: 'step' = 'step';
-    /** alternate between vertical and horizontal segments, as in a step function */
-    export const STEP_BEFORE: 'step-before' = 'step-before';
-    /** alternate between horizontal and vertical segments, as in a step function */
-    export const STEP_AFTER: 'step-after' = 'step-after';
-    /** a B-spline, with control point duplication on the ends */
-    export const BASIS: 'basis' = 'basis';
-    /** an open B-spline; may not intersect the start or end */
-    export const BASIS_OPEN: 'basis-open' = 'basis-open';
-    /** a closed B-spline, as in a loop */
-    export const BASIS_CLOSED: 'basis-closed' = 'basis-closed';
-    /** a Cardinal spline, with control point duplication on the ends */
-    export const CARDINAL: 'cardinal' = 'cardinal';
-    /** an open Cardinal spline; may not intersect the start or end, but will intersect other control points */
-    export const CARDINAL_OPEN: 'cardinal-open' = 'cardinal-open';
-    /** a closed Cardinal spline, as in a loop */
-    export const CARDINAL_CLOSED: 'cardinal-closed' = 'cardinal-closed';
-    /** equivalent to basis, except the tension parameter is used to straighten the spline */
-    export const BUNDLE: 'bundle' = 'bundle';
-    /** cubic interpolation that preserves monotonicity in y */
-    export const MONOTONE: 'monotone' = 'monotone';
-}
-export type Interpolate = typeof Interpolate.BASIS | typeof Interpolate.BASIS_CLOSED | typeof Interpolate.BASIS_OPEN
-  | typeof Interpolate.BUNDLE | typeof Interpolate.CARDINAL | typeof Interpolate.CARDINAL_CLOSED | typeof Interpolate.CARDINAL_OPEN
-  | typeof Interpolate.LINEAR | typeof Interpolate.LINEAR_CLOSED | typeof Interpolate.MONOTONE | typeof Interpolate.STEP
-  | typeof Interpolate.STEP_AFTER | typeof Interpolate.STEP_BEFORE;
-
+// FIXME refactor this
 export namespace AreaOverlay {
   export const LINE: 'line' = 'line';
   export const LINEPOINT: 'linepoint' = 'linepoint';
@@ -182,229 +109,6 @@ export const defaultOverlayConfig: OverlayConfig = {
   pointStyle: {filled: true},
   lineStyle: {}
 };
-
-export interface MarkConfig {
-
-  // ---------- Color ----------
-  /**
-   * Whether the shape\'s color should be used as fill color instead of stroke color.
-   * This is only applicable for "bar", "point", and "area".
-   * All marks except "point" marks are filled by default.
-   * See Mark Documentation (http://vega.github.io/vega-lite/docs/marks.html)
-   * for usage example.
-   */
-  filled?: boolean;
-
-  /**
-   * Default color.
-   */
-  color?: string;
-
-  /**
-   * Default Fill Color.  This has higher precedence than config.color
-   */
-  fill?: string;
-
-  /**
-   * Default Stroke Color.  This has higher precedence than config.color
-   */
-  stroke?: string;
-
-  // ---------- Opacity ----------
-  /**
-   * @minimum 0
-   * @maximum 1
-   */
-  opacity?: number;
-
-  /**
-   * @minimum 0
-   * @maximum 1
-   */
-  fillOpacity?: number;
-
-  /**
-   * @minimum 0
-   * @maximum 1
-   */
-  strokeOpacity?: number;
-
-  // ---------- Stroke Style ----------
-  /**
-   * @minimum 0
-   */
-  strokeWidth?: number;
-
-  /**
-   * An array of alternating stroke, space lengths for creating dashed or dotted lines.
-   */
-  strokeDash?: number[];
-
-  /**
-   * The offset (in pixels) into which to begin drawing with the stroke dash array.
-   */
-  strokeDashOffset?: number;
-
-  // ---------- Stacking: Bar & Area ----------
-  stacked?: StackOffset;
-
-  // ---------- Orientation: Bar, Tick, Line, Area ----------
-  /**
-   * The orientation of a non-stacked bar, tick, area, and line charts.
-   * The value is either horizontal (default) or vertical.
-   * - For bar, rule and tick, this determines whether the size of the bar and tick
-   * should be applied to x or y dimension.
-   * - For area, this property determines the orient property of the Vega output.
-   * - For line, this property determines the sort order of the points in the line
-   * if `config.sortLineBy` is not specified.
-   * For stacked charts, this is always determined by the orientation of the stack;
-   * therefore explicitly specified value will be ignored.
-   */
-  orient?: Orient;
-
-  // ---------- Interpolation: Line / area ----------
-  /**
-   * The line interpolation method to use. One of linear, step-before, step-after, basis, basis-open, cardinal, cardinal-open, monotone.
-   */
-  interpolate?: Interpolate;
-  /**
-   * Depending on the interpolation type, sets the tension parameter.
-   */
-  tension?: number;
-
-  // ---------- Line ---------
-  /**
-   * Size of line mark.
-   */
-  lineSize?: number;
-
-  // ---------- Rule ---------
-  /**
-   * Size of rule mark.
-   */
-  ruleSize?: number;
-
-  // ---------- Bar ----------
-  /**
-   * The size of the bars.  If unspecified, the default size is  `bandSize-1`,
-   * which provides 1 pixel offset between bars.
-   */
-  barSize?: number;
-
-  /**
-   * The size of the bars on continuous scales.
-   */
-  barThinSize?: number;
-
-  // ---------- Point ----------
-  /**
-   * The symbol shape to use. One of circle (default), square, cross, diamond, triangle-up, or triangle-down, or a custom SVG path.
-   */
-  shape?: Shape | string;
-
-  // ---------- Point Size (Point / Square / Circle) ----------
-  /**
-   * The pixel area each the point. For example: in the case of circles, the radius is determined in part by the square root of the size value.
-   */
-  size?: number;
-
-  // ---------- Tick ----------
-  /** The width of the ticks. */
-  tickSize?: number;
-
-  /** Thickness of the tick mark. */
-  tickThickness?: number;
-
-  // ---------- Text ----------
-  /**
-   * The horizontal alignment of the text. One of left, right, center.
-   */
-  align?: HorizontalAlign;
-  /**
-   * The rotation angle of the text, in degrees.
-   */
-  angle?: number;
-  /**
-   * The vertical alignment of the text. One of top, middle, bottom.
-   */
-  baseline?: VerticalAlign;
-  /**
-   * The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.
-   */
-  dx?: number;
-  /**
-   * The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.
-   */
-  dy?: number;
-  /**
-   * Polar coordinate radial offset, in pixels, of the text label from the origin determined by the x and y properties.
-   */
-  radius?: number;
-  /**
-   * Polar coordinate angle, in radians, of the text label from the origin determined by the x and y properties. Values for theta follow the same convention of arc mark startAngle and endAngle properties: angles are measured in radians, with 0 indicating "north".
-   */
-  theta?: number;
-  /**
-   * The typeface to set the text in (e.g., Helvetica Neue).
-   */
-  font?: string;
-  /**
-   * The font size, in pixels.
-   */
-  fontSize?: number;
-  /**
-   * The font style (e.g., italic).
-   */
-  fontStyle?: FontStyle;
-  /**
-   * The font weight (e.g., `"normal"`, `"bold"`, `900`).
-   */
-  fontWeight?: FontWeight | number;
-  // Vega-Lite only for text only
-  /**
-   * The formatting pattern for text value. If not defined, this will be determined automatically.
-   */
-  format?: string;
-  /**
-   * Whether month names and weekday names should be abbreviated.
-   */
-  shortTimeLabels?: boolean;
-  /**
-   * Placeholder Text
-   */
-  text?: string;
-
-  /**
-   * Offset between bar for binned field
-   * @minimum 0
-   */
-  barBinSpacing?: number;
-
-  /**
-   * Apply color field to background color instead of the text.
-   */
-  applyColorToBackground?: boolean;
-}
-
-export const defaultMarkConfig: MarkConfig = {
-  color: '#4682b4',
-  shape: Shape.CIRCLE,
-  strokeWidth: 2,
-  size: 30,
-  barThinSize: 2,
-  // lineSize is undefined by default, and refer to value from strokeWidth
-  ruleSize: 1,
-  tickThickness: 1,
-
-  fontSize: 10,
-  baseline: VerticalAlign.MIDDLE,
-  text: 'Abc',
-
-  barBinSpacing: 1,
-
-  applyColorToBackground: false
-};
-
 
 export interface Config {
   // TODO: add this back once we have top-down layout approach
