+++ src/mark.ts
@@ -1,3 +1,5 @@
+import {StackOffset} from './stack';
+
 export namespace Mark {
   export const AREA: 'area' = 'area';
   export const BAR: 'bar' = 'bar';
@@ -27,3 +29,251 @@ export const SQUARE = Mark.SQUARE;
 
 export const ERRORBAR = Mark.ERRORBAR;
 export const PRIMITIVE_MARKS = [AREA, BAR, LINE, POINT, TEXT, TICK, RULE, CIRCLE, SQUARE];
+
+export type FontStyle = 'normal' | 'italic';
+export type FontWeight = 'normal' | 'bold';
+export type HorizontalAlign = 'left' | 'right' | 'center';
+export type Interpolate = 'linear' | 'linear-closed' |
+  'step' | 'step-before' | 'step-after' |
+  'basis' | 'basis-open' | 'basis-closed' |
+  'cardinal' | 'cardinal-open' | 'cardinal-closed' |
+  'bundle' | 'monotone';
+export type Shape = 'circle' | 'square' | 'cross' | 'diamond' | 'triangle-up' | 'triangle-down';
+export type Orient = 'horizontal' | 'vertical';
+export type VerticalAlign = 'top' | 'middle' | 'bottom';
+
+export interface MarkConfig {
+
+  // ---------- Color ----------
+  /**
+   * Whether the shape\'s color should be used as fill color instead of stroke color.
+   * This is only applicable for "bar", "point", and "area".
+   * All marks except "point" marks are filled by default.
+   * See Mark Documentation (http://vega.github.io/vega-lite/docs/marks.html)
+   * for usage example.
+   */
+  filled?: boolean;
+
+  /**
+   * Default color.
+   */
+  color?: string;
+
+  /**
+   * Default Fill Color.  This has higher precedence than config.color
+   */
+  fill?: string;
+
+  /**
+   * Default Stroke Color.  This has higher precedence than config.color
+   */
+  stroke?: string;
+
+  // ---------- Opacity ----------
+  /**
+   * @minimum 0
+   * @maximum 1
+   */
+  opacity?: number;
+
+  /**
+   * @minimum 0
+   * @maximum 1
+   */
+  fillOpacity?: number;
+
+  /**
+   * @minimum 0
+   * @maximum 1
+   */
+  strokeOpacity?: number;
+
+  // ---------- Stroke Style ----------
+  /**
+   * @minimum 0
+   */
+  strokeWidth?: number;
+
+  /**
+   * An array of alternating stroke, space lengths for creating dashed or dotted lines.
+   */
+  strokeDash?: number[];
+
+  /**
+   * The offset (in pixels) into which to begin drawing with the stroke dash array.
+   */
+  strokeDashOffset?: number;
+
+  // FIXME: move this outside mark config
+  // ---------- Stacking: Bar & Area ----------
+  stacked?: StackOffset;
+
+  // ---------- Orientation: Bar, Tick, Line, Area ----------
+  /**
+   * The orientation of a non-stacked bar, tick, area, and line charts.
+   * The value is either horizontal (default) or vertical.
+   * - For bar, rule and tick, this determines whether the size of the bar and tick
+   * should be applied to x or y dimension.
+   * - For area, this property determines the orient property of the Vega output.
+   * - For line, this property determines the sort order of the points in the line
+   * if `config.sortLineBy` is not specified.
+   * For stacked charts, this is always determined by the orientation of the stack;
+   * therefore explicitly specified value will be ignored.
+   */
+  orient?: Orient;
+
+  // ---------- Interpolation: Line / area ----------
+  /**
+   * The line interpolation method to use. One of the following:
+   * - `"linear"`: piecewise linear segments, as in a polyline.
+   * - `"linear-closed"`: close the linear segments to form a polygon.
+   * - `"step"`: alternate between horizontal and vertical segments, as in a step function.
+   * - `"step-before"`: alternate between vertical and horizontal segments, as in a step function.
+   * - `"step-after"`: alternate between horizontal and vertical segments, as in a step function.
+   * - `"basis"`: a B-spline, with control point duplication on the ends.
+   * - `"basis-open"`: an open B-spline; may not intersect the start or end.
+   * - `"basis-closed"`: a closed B-spline, as in a loop.
+   * - `"cardinal"`: a Cardinal spline, with control point duplication on the ends.
+   * - `"cardinal-open"`: an open Cardinal spline; may not intersect the start or end, but will intersect other control points.
+   * - `"cardinal-closed"`: a closed Cardinal spline, as in a loop.
+   * - `"bundle"`: equivalent to basis, except the tension parameter is used to straighten the spline.
+   * - `"monotone"`: cubic interpolation that preserves monotonicity in y.
+   */
+  interpolate?: Interpolate;
+  /**
+   * Depending on the interpolation type, sets the tension parameter.
+   */
+  tension?: number;
+
+  // ---------- Line ---------
+  /**
+   * Size of line mark.
+   */
+  lineSize?: number;
+
+  // ---------- Rule ---------
+  /**
+   * Size of rule mark.
+   */
+  ruleSize?: number;
+
+  // ---------- Bar ----------
+  /**
+   * The size of the bars.  If unspecified, the default size is  `bandSize-1`,
+   * which provides 1 pixel offset between bars.
+   */
+  barSize?: number;
+
+  /**
+   * The size of the bars on continuous scales.
+   */
+  barThinSize?: number;
+
+  // ---------- Point ----------
+  /**
+   * The symbol shape to use. One of circle (default), square, cross, diamond, triangle-up, or triangle-down, or a custom SVG path.
+   */
+  shape?: Shape | string;
+
+  // ---------- Point Size (Point / Square / Circle) ----------
+  /**
+   * The pixel area each the point. For example: in the case of circles, the radius is determined in part by the square root of the size value.
+   */
+  size?: number;
+
+  // ---------- Tick ----------
+  /** The width of the ticks. */
+  tickSize?: number;
+
+  /** Thickness of the tick mark. */
+  tickThickness?: number;
+
+  // ---------- Text ----------
+  /**
+   * The horizontal alignment of the text. One of left, right, center.
+   */
+  align?: HorizontalAlign;
+  /**
+   * The rotation angle of the text, in degrees.
+   */
+  angle?: number;
+  /**
+   * The vertical alignment of the text. One of top, middle, bottom.
+   */
+  baseline?: VerticalAlign;
+  /**
+   * The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.
+   */
+  dx?: number;
+  /**
+   * The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.
+   */
+  dy?: number;
+  /**
+   * Polar coordinate radial offset, in pixels, of the text label from the origin determined by the x and y properties.
+   */
+  radius?: number;
+  /**
+   * Polar coordinate angle, in radians, of the text label from the origin determined by the x and y properties. Values for theta follow the same convention of arc mark startAngle and endAngle properties: angles are measured in radians, with 0 indicating "north".
+   */
+  theta?: number;
+  /**
+   * The typeface to set the text in (e.g., Helvetica Neue).
+   */
+  font?: string;
+  /**
+   * The font size, in pixels.
+   */
+  fontSize?: number;
+  /**
+   * The font style (e.g., italic).
+   */
+  fontStyle?: FontStyle;
+  /**
+   * The font weight (e.g., `"normal"`, `"bold"`, `900`).
+   */
+  fontWeight?: FontWeight | number;
+  // Vega-Lite only for text only
+  /**
+   * The formatting pattern for text value. If not defined, this will be determined automatically.
+   */
+  format?: string;
+  /**
+   * Whether month names and weekday names should be abbreviated.
+   */
+  shortTimeLabels?: boolean;
+  /**
+   * Placeholder Text
+   */
+  text?: string;
+
+  /**
+   * Offset between bar for binned field
+   * @minimum 0
+   */
+  barBinSpacing?: number;
+
+  /**
+   * Apply color field to background color instead of the text.
+   */
+  applyColorToBackground?: boolean;
+}
+
+export const defaultMarkConfig: MarkConfig = {
+  color: '#4682b4',
+  shape: 'circle',
+  strokeWidth: 2,
+  size: 30,
+  barThinSize: 2,
+  // lineSize is undefined by default, and refer to value from strokeWidth
+  ruleSize: 1,
+  tickThickness: 1,
+
+  fontSize: 10,
+  baseline: 'middle',
+  text: 'Abc',
+
+  barBinSpacing: 1,
+
+  applyColorToBackground: false
+};
